pypy enhancements
