+++ FastAPI/offers/tests/conftest.py
@@ -307,7 +307,7 @@ def offer_ps4(
 
 
 @pytest.fixture
-def chat_marian_egor(db, user_marian, user_egor):
+def chat_marian_egor(db, user_marian, user_egor, offer_ps4):
     query = """
     INSERT INTO chats (
         chat_id,
@@ -320,12 +320,12 @@ def chat_marian_egor(db, user_marian, user_egor):
     RETURNING offer_id
     """
     values = (
-        "20",  # chat_id
-        "1",  # author_id
-        "2",  # client_id
-        "a30b8a1e-1c60-4bbc-ac3d-37df2d224000",  # offer_id
+        "1",  # chat_id
+        user_egor,  # author_id
+        user_egor,  # client_id
+        offer_ps4,  # offer_id
         "2021-06-21 07:52:21.609079+00",  # creation_datetime
-        False  # is_done
+        False,  # is_done
     )
     db.execute(query, values)
-    return "20"
+    return "1"
